Continue on RabbitMQ refactoring
